+++ application/controllers/Details/PartDetails.php
@@ -9,16 +9,23 @@ class PartDetails extends Application
 		parent::__construct();
 		$this->load->model('PartsModel');
 	}
-	public function index()
-	{
-                          
-	}
 
     public function get($id){
-        $data = $this->PartsModel->get($id);
-        $thid->data['part'] =  $data;
-        $this->data['pagebody'] ='part_details';   
+		
+        $role = $this->session->userdata('userrole');
 
-        $this->render(); 
+        if(strtolower($role) == 'boss' || strtolower($role) == 'supervisor' || strtolower($role) == 'worker'){
+
+			$parts = $this->PartsModel->get($id);
+			$parts[0]['img'] = $parts[0]['part_code'] . '.jpeg';
+			$this->data['parts'] = $parts;
+
+			$this->data['pagebody'] ='part_details';   
+
+			$this->render(); 
+		}       
+        else{
+            redirect('Welcome');
+        }
     }
 }